+++ CHANGELOG.md
@@ -26,6 +26,9 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Added
 
+- Confirmation-protected clear actions on the Query Log and Dashboard pages
+  ([#3896], [#4087]).
+
 - Bootstrap servers configuration now supports comments.
 
 - New property `"language"` in `POST /control/install/check_config` and `POST /control/install/configure` HTTP APIs.
@@ -48,6 +51,8 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 [#7514]: https://github.com/AdguardTeam/AdGuardHome/issues/7514
 [#8183]: https://github.com/AdguardTeam/AdGuardHome/issues/8183
+[#3896]: https://github.com/AdguardTeam/AdGuardHome/issues/3896
+[#4087]: https://github.com/AdguardTeam/AdGuardHome/issues/4087
 
 <!--
 NOTE: Add new changes ABOVE THIS COMMENT.